fix: Внесены коррективы в ER - диаграмму, описание БД. feat: Объекту Film добавлено новое свойство - rate. Добавлены классы и интерфейсы- хранилища для FilmAudience, FilmGenre, Friend, а также реализована логика этих классов

+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmDbStorage.java
@@ -4,7 +4,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
@@ -16,8 +15,9 @@ import ru.yandex.practicum.filmorate.storage.genre.GenreMapper;
 
 import java.sql.Date;
 import java.sql.PreparedStatement;
-import java.sql.SQLException;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.List;
 
 @Slf4j
 @Repository
@@ -28,29 +28,35 @@ public class FilmDbStorage implements FilmStorage {
 
     @Override
     public Film addObject(Film object) {
-        String sql = "INSERT INTO films(name, description, release_date, duration, mpa_id) " +
-                "VALUES (?, ?, ?, ?, ?)";
+        String sql = "INSERT INTO films(name, description, release_date, duration, rate = ?, mpa_id) " +
+                "VALUES (?, ?, ?, ?, ?, ?)";
         KeyHolder keyHolder = new GeneratedKeyHolder();
         jdbcTemplate.update(connection -> {
             PreparedStatement stmt = connection.prepareStatement(sql, new String[]{"film_id"});
             stmt.setString(1, object.getName());
             stmt.setString(2, object.getDescription());
             stmt.setDate(3, Date.valueOf(object.getReleaseDate()));
-            stmt.setLong(4, object.getDuration());
-            stmt.setLong(5, object.getMpa().getId());
+            stmt.setInt(4, object.getDuration());
+            stmt.setLong(5, object.getRate());
+            stmt.setLong(6, object.getMpa().getId());
             return stmt;
         }, keyHolder);
         object.setId(keyHolder.getKey().longValue());
         saveGenresToFilm(object);
+        log.info(LogMessages.OBJECT_ADDED.toString(), object);
         return object;
     }
 
     @Override
     public Film renewalObject(Film object) {
         String sql = "UPDATE film_data SET name = ?, description = ?, release_date = ?, " +
-                "duration = ?, mpa_id = ? where film_id = ?";
-        jdbcTemplate.update(sql, object.getName(), object.getDescription(), object.getReleaseDate(),
-                object.getDuration(), object.getMpa().getId(), object.getId());
+                "duration = ?, rate = ?, mpa_id = ? where film_id = ?";
+        int filmData = jdbcTemplate.update(sql, object.getName(), object.getDescription(), object.getReleaseDate(),
+                object.getDuration(), object.getRate(), object.getMpa().getId(), object.getId());
+        if(filmData == 0) {
+            log.warn(LogMessages.OBJECT_NOT_FOUND.toString());
+            throw new ObjectNotFoundException(LogMessages.OBJECT_NOT_FOUND.toString());
+        }
         jdbcTemplate.update("DELETE FROM film_genre WHERE film_id = ?", object.getId());
         if (object.getGenres() != null && !object.getGenres().isEmpty()) {
             for (Genre genre : object.getGenres()) {
@@ -74,17 +80,17 @@ public class FilmDbStorage implements FilmStorage {
                 "INNER JOIN mpa ON film_data.mpa_id = mpa.mpa_id\n" +
                 "WHERE film_data.film_id = ?;";
         Film film = jdbcTemplate.queryForObject(sql, new FilmMapper(), id);
-        String genresSql = "SELECT *\n" +
+        String genreSql= "SELECT *\n" +
                 "FROM genre\n" +
                 "INNER JOIN film_genre AS fg ON film_data.film_id = fg.film_id\n" +
                 "INNER JOIN film_data ON fg.film_id = film_data.film_id \n" +
                 "WHERE film_data.film_id = ?";
-        List<Genre> genres = jdbcTemplate.query(genresSql, new GenreMapper(), id);
+        List<Genre> genre = jdbcTemplate.query(genreSql, new GenreMapper(), id);
         if (film != null) {
-            film.setGenres(genres);
+            film.setGenres(genre);
         } else {
-            log.warn(LogMessages.NULL_OBJECT.toString());
-            throw new ObjectNotFoundException(LogMessages.NULL_OBJECT.toString());
+            log.warn(LogMessages.OBJECT_NOT_FOUND.toString());
+            throw new ObjectNotFoundException(LogMessages.OBJECT_NOT_FOUND.toString());
         }
         return film;
     }
@@ -100,25 +106,6 @@ public class FilmDbStorage implements FilmStorage {
                 "INNER JOIN genre ON fg.genre_id = genre.genre_id";
         return jdbcTemplate.query(sql, new FilmMapper());
     }
-
-//    private boolean dbContainsFilm(Film film) {
-//        String sql = "SELECT fd.*, mpa.name \n" +
-//                "FROM film_data AS fd \n" +
-//                "JOIN mpa ON fd.mpa_id = mpa.mpa_id \n" +
-//                "WHERE fd.name = ? \n" +
-//                "AND  fd.description = ?\n" +
-//                "AND fd.release_date = ? \n" +
-//                "AND fd.duration = ?\n" +
-//                "AND fd.mpa_id = ?";
-//        Film f = jdbcTemplate.queryForObject(sql, new FilmMapper(), film.getName(), film.getDescription(),
-//                film.getReleaseDate(), film.getDuration(), film.getMpa().getId());
-//        if (f == null) {
-//            log.warn(LogMessages.NULL_OBJECT.toString());
-//            return false;
-//        } else {
-//            return true;
-//        }
-//    }
 
     private void saveGenresToFilm(Film film) {
         long filmId = film.getId();
